# Improve performance of unpack() with nameless repetitions