@staabm 2025-06-08
nice to see the benchmark was useful. thanks for looking into it.