@nielsdos 2025-06-08
Looks like there's a missing null check somewhere, I'll look tomorrow or this evening.